requirements.txt
google-generativeai
streamlit
ipython
python-dotenv
pillow